refactor(mastering): extract Stages 7+7.5; wire thin orchestrator (#290 D5)
master_album is now a ~75-line orchestrator calling 14 stage functions
in _album_stages.py. Zero behavior change — all existing tests pass.
ADM validation (step 9) can now be added as _stage_adm_validation
without making audio.py harder to read.
